new tonic_finance.db

+++ models.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from sqlalchemy import create_engine, Column, Integer, String, Float, DateTime, ForeignKey
 from sqlalchemy.orm import declarative_base, relationship
 
@@ -10,7 +9,7 @@ class User(base):
     username = Column(String(50), nullable=False)
     email = Column(String(100), unique=True, nullable=False)
     password = Column(String(225), nullable=False)
-    initial_balance = Column(Decimal, default=0.0)
+    initial_balance = Column(Float, default=0.0)
     transactions = relationship("Transactions", back_populates="user", cascade="all, delete-orphan")
     def __repr__(self):
         return f"<User(username='{self.username}', email='{self.email}')>"
@@ -19,7 +18,7 @@ class Transactions(base):
     id = Column(Integer, primary_key=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     type = Column(String(10), nullable=False)
-    amount = Column(Decimal, nullable=False)
+    amount = Column(Float, nullable=False)
     category = Column(String(50), nullable=False)
     description = Column(String(225))
     date = Column(DateTime, default=datetime.utcnow)
